About to migrate all the hashmap utilities to a module 'hashmap_utilities.f90' to keep them all general and reuseable, rather than copying and pasting almost the same code everywhere.

+++ common/stochastics.f90
@@ -239,7 +239,7 @@ end function get_covariance
 
 pure function hill(x, k, n, c) result(f)
 ! Hill function. n<0 for negative hill function, n>0 for positive.
-	real, intent(in) :: x
+	real(dp), intent(in) :: x
 	real(dp), intent(in) :: k, n, c
 	real(dp) :: f
